requirements.txt
# HTTP请求库
requests>=2.28.0

# 数据处理
pandas>=1.5.0
numpy>=1.22.0

# 进度条显示
tqdm>=4.64.0

# 环境变量管理
python-dotenv>=0.20.0

# E2B SDK
e2b-code-interpreter>=0.5.0

# 统计计算
statistics>=1.0.3 
